the basics of RabbitMQ and creating a producer and consumer for a single queue.
