+++ RobbitMQ.Consumer/Program.cs
@@ -1,7 +1,5 @@
-﻿using Newtonsoft.Json;
-using RabbitMQ.Client;
-using RabbitMQ.Client.Events;
-using System.Text;
+﻿using RabbitMQ.Client;
+using RobbitMQ.Consumer;
 
 var factory = new ConnectionFactory
 {
@@ -10,18 +8,4 @@ var factory = new ConnectionFactory
 
 using var connection = factory.CreateConnection();  // Corrected: Use CreateConnection() directly
 using var channel = connection.CreateModel();
-
-channel.QueueDeclare("demo-queue",
-    durable: true,
-    exclusive: false,
-    autoDelete: false,
-    arguments: null);  // Corrected: Removed invalid parameter
-var consumer = new EventingBasicConsumer(channel);
-consumer.Received += (sender, eventArgs) =>
-{
-    var body = eventArgs.Body.ToArray();
-    var message = Encoding.UTF8.GetString(body);
-    Console.WriteLine($"Received: {message}");
-};
-channel.BasicConsume("demo-queue", true,consumer);
-Console.ReadLine();
+DirectExchangeConsumer.Consume(channel);
